page layout, site index and show page modification, and welcome page with route function modified
created a page layout in the component folder for the overall general structure of the sites, inspections, and reports. Then modified the index and show of the sites folder to remove the html doc structure content and simply have the page layout tag for referencing it. then moved the route welcome function to the top and now the welcome file has the page layout tag inside it as the main welcome page